gallehr/cbam#561: updated cn_code and Raw Mass [t]
